refactor global kernel/grad registries
